Move create function action to app service lib
